+++ tests/it/io/parquet/mod.rs
@@ -6,8 +6,8 @@ use arrow2::{
     io::parquet::read::statistics::*, io::parquet::read::*, io::parquet::write::*,
 };
 
-use crate::io::ipc::read_gzip_json;
-
+#[cfg(feature = "io_json_integration")]
+mod integration;
 mod read;
 mod read_indexes;
 mod write;
@@ -789,7 +789,6 @@ pub fn pyarrow_struct_statistics(column: &str) -> Statistics {
     }
 }
 
-/// Round-trip with parquet using the same integration files used for IPC integration tests.
 fn integration_write(schema: &Schema, batches: &[Chunk<Arc<dyn Array>>]) -> Result<Vec<u8>> {
     let options = WriteOptions {
         write_statistics: true,
@@ -839,43 +838,6 @@ fn integration_read(data: &[u8]) -> Result<IntegrationRead> {
     let batches = reader.collect::<Result<Vec<_>>>()?;
 
     Ok((schema, batches))
-}
-
-fn test_file(version: &str, file_name: &str) -> Result<()> {
-    let (schema, _, batches) = read_gzip_json(version, file_name)?;
-
-    // empty batches are not written/read from parquet and can be ignored
-    let batches = batches
-        .into_iter()
-        .filter(|x| !x.is_empty())
-        .collect::<Vec<_>>();
-
-    let data = integration_write(&schema, &batches)?;
-
-    let (read_schema, read_batches) = integration_read(&data)?;
-
-    assert_eq!(schema, read_schema);
-    assert_eq!(batches, read_batches);
-
-    Ok(())
-}
-
-#[test]
-fn roundtrip_100_primitive() -> Result<()> {
-    test_file("1.0.0-littleendian", "generated_primitive")?;
-    test_file("1.0.0-bigendian", "generated_primitive")
-}
-
-#[test]
-fn roundtrip_100_dict() -> Result<()> {
-    test_file("1.0.0-littleendian", "generated_dictionary")?;
-    test_file("1.0.0-bigendian", "generated_dictionary")
-}
-
-#[test]
-fn roundtrip_100_extension() -> Result<()> {
-    test_file("1.0.0-littleendian", "generated_extension")?;
-    test_file("1.0.0-bigendian", "generated_extension")
 }
 
 /// Tests that when arrow-specific types (Duration and LargeUtf8) are written to parquet, we can rountrip its
